requirements.txt
opencv-python
opencv-contrib-python
mysql-connector-python
Pillow
customtkinter
bcrypt
watchdog
tksheet
